allow user to select file batches for download

+++ scripts/01-target-download.sh
@@ -2,8 +2,8 @@
 
 # Download script for TARGET bulk RNA-seq for pediatric cancer splicing analysis
 # this script downloads TARGET cancer datasets obtained from filter_target_sra.qmd, which includes all TARGET paired-end RNA-seq samples, excluding ssRNA-seq samples, cell lines and xenografts, and samples known to be ribo-deplete
-# usage: scripts/01-target-download.sh [gtex|target]
-# example for downloading target: scripts/01-target-download.sh target
+# usage: scripts/01-target-download.sh [gtex|target] [batch number]
+# example for downloading the first 1.2TB batch of target: scripts/01-target-download.sh target 1
 
 # cause nonzero exit status and undefined variables to stop the script
 set -euo pipefail
@@ -53,42 +53,48 @@ exec > >(tee $log_file) 2>&1
 # Run filtering script to obtain accession IDs to download
 quarto render ${target_filtering}
 
-# Use accession IDs in first column of accessions file to download fastqs for analysis
-# iterate through first column of TARGET accession metadata, which has accession IDs
-# skip the first line, which consists of column headers
-time for ID in $(awk 'NR>1{print $1}' "${target_gtex_dir}/$1_accessions.tsv"); do
-    # check that fastq or zipped fastq does not already exist
-    if [ ! -f "${fastq_dir}/${ID}_1.fastq" ] && [ ! -f "${fastq_dir}/${ID}_1.fastq.gz" ]; then
-        # prefetch file dependencies
-        # set max size of prefetch file to 50GB based on max file size of samples in accessions lists
-        # continue even if an error is thrown
-        prefetch --ngc $dbgap_key --max-size 50000000 --output-directory $fastq_dir $ID || true
-        # Check if prefetch file directory has been created
-        # Lack of directory means prefetch has failed; skip to next accession
-        if [ ! -d "${fastq_dir}/$ID" ]; then
-            # record the accession ID not downloaded
-            echo -e "$ID\tprefetch" >> $error_log_file
-            continue
+# Select 1.2TB batch of accessions to download based on user input
+# batch ID is in the last column of the metadata
+if batch in $(awk 'NF>1{print $1}' "${metadata_dir}/$1_accessions.tsv") == ${2}; then
+
+    # Use accession IDs in first column of accessions file to download fastqs for analysis
+    # iterate through first column of TARGET accession metadata, which has accession IDs
+    # skip the first line, which consists of column headers
+    time for ID in $(awk 'NR>1{print $1}' "${metadata_dir}/$1_accessions.tsv"); do
+        # check that fastq or zipped fastq does not already exist
+        if [ ! -f "${fastq_dir}/${ID}_1.fastq" ] && [ ! -f "${fastq_dir}/${ID}_1.fastq.gz" ]; then
+            # prefetch file dependencies
+            # set max size of prefetch file to 50GB based on max file size of samples in accessions lists
+            # continue even if an error is thrown
+            prefetch --ngc $dbgap_key --max-size 50000000 --output-directory $fastq_dir $ID || true
+            # Check if prefetch file directory has been created
+            # Lack of directory means prefetch has failed; skip to next accession
+            if [ ! -d "${fastq_dir}/$ID" ]; then
+                # record the accession ID not downloaded
+                echo -e "$ID\tprefetch" >> $error_log_file
+                continue
+            fi
+
+            # download fastq file for each accession
+            fasterq-dump --temp $group_dir --ngc $dbgap_key $ID --threads 4 --outdir $fastq_dir || true
+
+            # Check if fastq file exists
+            # If file is missing, record error in fastq-dump step and move on to next accession
+            if [ ! -f "${fastq_dir}"/${ID}_1.fastq || ! -f "${fastq_dir}"/${ID}_2.fastq]; then
+                # record the accession ID not downloaded
+                echo -e "$ID\tfasterq-dump" >> $error_log_file
+                continue
+            fi
+
+            # zip fastqs to save space
+            # use default number of processes, which is 8 or number of online processors
+            pigz "${fastq_dir}"/${ID}_*.fastq
+
+            # offload prefetch prerequisites
+            rm -R "${fastq_dir}/${ID}"
+
         fi
 
-        # download fastq file for each accession
-        fasterq-dump --temp $group_dir --ngc $dbgap_key $ID --threads 4 --outdir $fastq_dir || true
+    done
 
-        # Check if fastq file exists
-        # If file is missing, record error in fastq-dump step and move on to next accession
-        if [ ! -f "${fastq_dir}"/${ID}_1.fastq || ! -f "${fastq_dir}"/${ID}_2.fastq]; then
-            # record the accession ID not downloaded
-            echo -e "$ID\tfasterq-dump" >> $error_log_file
-            continue
-        fi
-
-        # zip fastqs to save space
-        # use default number of processes, which is 8 or number of online processors
-        pigz "${fastq_dir}"/${ID}_*.fastq
-
-        # offload prefetch prerequisites
-        rm -R "${fastq_dir}/${ID}"
-
-    fi
-
-done
+fi
